rename: store.ts -> window.ts
