+++ src/__mocks__/window.ts
